+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog ##
 #### dev ####
+* No longer require patch classes and methods to be public
 
 #### v3.16.0 ####
 * Fix patch error with ByRef while mixing primitive and wrapper types